requirements.txt
django==1.2.1
django-socialregistration
